// vsu-app/ClientApp/src/model/invoiceLineItemDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * VSU API
 * API for the Victim Services Unit Application (VSU) application
 * OpenAPI spec version: v1
 */

export interface InvoiceLineItemDto {
  /** @nullable */
  expenseType?: number | null;
  /** @nullable */
  transportationType?: number | null;
  /** @nullable */
  mileage?: number | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  other?: string | null;
  /** @nullable */
  number?: number | null;
  /** @nullable */
  dailyRoomRate?: number | null;
  /** @nullable */
  childAge?: number | null;
  /** @nullable */
  childcareStartDate?: string | null;
  /** @nullable */
  childcareEndDate?: string | null;
  /**
   * @minLength 0
   * @maxLength 160
   * @nullable
   */
  childcareProviderFirstName?: string | null;
  /**
   * @minLength 0
   * @maxLength 160
   * @nullable
   */
  childcareProviderLastName?: string | null;
  /**
   * @minLength 0
   * @maxLength 20
   * @nullable
   */
  childcareProviderPhoneNo?: string | null;
}