// vsu-app/ClientApp/src/model/reimbursementCaseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * VSU API
 * API for the Victim Services Unit Application (VSU) application
 * OpenAPI spec version: v1
 */
import type { CaseDto } from './caseDto';
import type { DocumentDto } from './documentDto';
import type { InvoiceDto } from './invoiceDto';
import type { InvoiceLineItemDto } from './invoiceLineItemDto';
import type { TravelInfoDto } from './travelInfoDto';

export interface ReimbursementCaseDto {
  caseId?: CaseDto;
  /** @nullable */
  contactInfoComments?: string | null;
  invoice?: InvoiceDto;
  /** @nullable */
  travelInfoCollection?: TravelInfoDto[] | null;
  /** @nullable */
  transportationExpenseCollection?: InvoiceLineItemDto[] | null;
  /** @nullable */
  accommodationExpenseCollection?: InvoiceLineItemDto[] | null;
  /** @nullable */
  mealExpenseCollection?: InvoiceLineItemDto[] | null;
  /** @nullable */
  childcareExpenseCollection?: InvoiceLineItemDto[] | null;
  /** @nullable */
  otherExpenseCollection?: InvoiceLineItemDto[] | null;
  /** @nullable */
  documentCollection?: DocumentDto[] | null;
}